Created file for only clickable frame and ended API req when cancelled"
